requirements.txt
streamlit
plotly
pyyaml
pandas
numpy